Make gke-exec-auth-plugin build for Windows.
The tpm2.OpenTPM() function requires different arguments for Windows vs.
Linux. This change imitates the existing kubernetes approach (e.g.
https://github.com/kubernetes/kubernetes/blob/master/pkg/util/mount/BUILD#L53) for building code on both Linux and Windows by splitting the openTPM() function implementation into OS-specific files.

This Windows build command now succeeds:

bazel build --platforms=@io_bazel_rules_go//go/toolchain:windows_amd64 \
  cmd/gke-exec-auth-plugin:gke-exec-auth-plugin
